refactor: organize TUI source by views
