+++ src/tui/utils/entry-utils.js
