+++ src/tui/utils/line-formatter.js
