+++ src/tui/utils/trace-buffer.js
