+++ src/tui/utils/trace-paths.js
