+++ src/tui/views/log-viewer/LineRenderer.js
