+++ test/tui/read-traces.test.js
@@ -7,7 +7,7 @@ import {
   readTraceEntries,
   listSessions,
   getTraceFile,
-} from '../../src/tui/read-traces.js';
+} from '../../src/tui/utils/read-traces.js';
 
 describe('read-traces helpers', () => {
   let tmpHome;
